Set a conversion rule

This function constructs the conversion between from one to the other land type.

Usage

clue_conversion(
  scene,
  from,
  to,
  label = NULL,
  thresh = NULL,
  auto = NULL,
  mask = NULL
)

Arguments

scene

description

from

character(1)
a land type from which a conversion shall be created.

to

character(1)
a land type to which the conversion shall be created.

label

character(1)
description of the conversion (used in the sequence plot).

thresh

integerish(1)
the number of time steps the land type in from must remain the same before it can change into the land type in to.

auto

integerish(1)
the number of time steps after which the land type in from will automatically change into to.

mask

character(1)
the name of a driver (see clue_driver) for which conversion to the specified class is allowed.