Set land types and their attributes

Set land types and their attributes

Usage

clue_landtype(
  scene,
  name,
  resistance = NULL,
  suitability = NULL,
  production = NULL,
  conversion = NULL,
  neighborhood = NULL,
  preference = NULL
)

Arguments

scene

description

name

character(1)
name of the land type.

resistance

numeric(1)
the relative conversion resistance to land change.

suitability

list(.)
list of the result of regression analysis. Includes the constant (beta_0) and as many drivers and their beta coefficient as there are (beta_1, beta_2, ...). Driver names must correspond to the names defined in clue_gridded. Each land type must have at least one suitability driver (with const = 1).

production

list(.)
optionally, list of services produced by this land type. Each list-item is a list itself, its name the service produced and the values of the list specify the priority and amount of production.

conversion

list(.)
optionally, list of land types into which this land type can be converted. Each list-item is a list itself, its name the land type into which the focal land type can be converted and the values of the list specify the details of conversion. Must have a value for 'label' and optionally one of min, max,auto or 'mask'.

neighborhood

list(1)
work in progress

preference

list(2)
work in progress