Set options of the current model run

Set options of the current model run

Usage

clue_options(
  iter_mode = NULL,
  conv1 = NULL,
  conv2 = NULL,
  out_type = NULL,
  rs_reg = NULL,
  ls_hist = NULL,
  neigh = NULL,
  loc_pref = NULL,
  dyn_lusmat = NULL,
  out_write = NULL,
  iter_param = NULL,
  protected = TRUE
)

Arguments

iter_mode

integerish(1)
Iteration mode:

1 = convergence criteria are expressed as absolute values,

2 = not yet implemented

conv1

numeric(1)
First convergence criterion.

conv2

numeric(1)
Second convergence criterion.

out_type

integerish(1)
Choice for the type of output file:

0 = no headers in output file,

1 = ArcGIS headers,

-2 = no headers and suppress information in log-file,

2 = ArcGIS headers and suppress information in log-file,

3 = ArcGIS/ArcMap extension and headers (default)

rs_reg

integerish(1)
Region specific regressions.

0 = no different regressions for different regions (default),

1 = different regressions for different regions,

2 = different regressions for different regions.

ls_hist

integerish(2)
Initialization of land use history:

0 = read from file age,

1 = generate random number + standard seed (default),

2 = merely random number

For options 1 and 2, an additional value for the maximum number of years of the random numbers must be given.

neigh

integerish(1)
Neighbourhood function (not yet implemented):

0 = off (default),

1 = on in simulation (automatically chosen on use of setNeighbourhood),

2 = only influences calculated, no simulation

loc_pref

integerish(1)

dyn_lusmat

integerish(1)
Dynamic landuse matrix (not yet implemented):

0 = off (default),

1 = on

out_write

integerish(1)
Write selected output measures:

0 = off,

1 = on for final year (default),

2 = on for all years

iter_param

numeric(1)
Choice for stability vs. speed of convergence. Value should be between 0.001 and 0.1, with higher values leading to more stable and more likely solution and lower values leading to faster and less likely solution.

protected

logical(1)
whether or not to consider protected areas (if the gridded object 'protected' has been given). (not implemented yet)

Value

no return value, called for the side-effect of setting model options.