Create a new CLUE scene

Create a new CLUE scene

Usage

clue_scene(root, name, period, regions, description = NULL)

Arguments

root

character(1)
path where the model files will be stored.

name

character(1)
the name of the scene.

period

integerish(2)
start and end year of the scene.

regions

character(1)
path to a geopackage file that has regions for which to apply a CLUE scene. Can optionally also be an opbject of class st_sf representation of that file.

description

character(1)
the description of the scene.

Value

object of class scene that contains initial meta data.

Examples

minimal <- clue_scene(
    root = paste0(tempdir(), "/ls2d1"),
    name = "ls2d1",
    period = c(2000, 2024),
    regions = paste0(system.file("test_data", package = "clueRIO"),
                    "/test_region.gpkg"),
    description = "this is a minimal showcase where forest is converted
                  to cropland to produce crops.")
#> creating model directory /tmp/RtmpFARxgK/ls2d1/.