Crop rotation modifier

Crop rotation modifier

Usage

mod_crop_rotation()

Details

To implement crop rotations, 'locspec' files are adapted for each year. The yearly 'locspec' is calculated by

selecting only a subset of pixels, depending on the sign of this weight factor. In case the sign is negative, it would keep those pixels that were covered by the focal class in the previous time-step and thus down-weigh where is was present previously. In case the sign is positive, it would keep those pixels that were covered by all other classes in the previous time-step and thus up-weigh where it was not present previously.