Specify neighbourhood models

This function sets the parameters for neigbourhood modelling. This requires to specify the neighbourhood (dimensions and weights) and a regression equation that determines how neighborhood results contribute to the suitability of a location.

Usage

setNeighbourhood(
  landuse = NULL,
  const = NULL,
  ...,
  weight = NULL,
  kernel = matrix(1, 3, 3),
  table = NULL
)

Arguments

landuse

character(1)
the landuse type for which to specify the neighbourhood parameters.

const

numeric(1)
constant of regression equation (how neighborhood results contribute to the suitability of a location) for this landuse type.

...

numeric(1)
named value, where the name is the landuse type in the neighbourhood and the value is the respective beta coefficient of the regression equation (how neighborhood results contribute to the suitability of a location).

weight

numeric(1)
the weight (0 - 1) assigned to the neighborhood function.

kernel

matrix(1)
the region within which neighbourhood shall be considered. This is a matrix with an odd number of columns and rows containing weights (0 - 1) of the cells in the neighbourhood.

table

data.frame(1)
object that contains one or more landuse types in rows and the columns "landuse" and "const" of each type and any number of landuse types with the respective model coefficients. (not yet implemented)

Details

When a land use type is not influenced by a neighborhood function (while others are), const is set to 0 and nothing else is specified.

See also