Set production of commodities by land-use types

This function registers how much of a good or service (commodity) is produced by land-use type(s).

Usage

setProduction(
  of = NULL,
  by = NULL,
  amount = NULL,
  priority = NULL,
  exclude = NULL
)

Arguments

of

character(1)
a commoditiy for which a production amount should be set.

by

character(.)
land-use type(s) by which of is produced.

amount

numeric(.)
the amount(s) of produced (must have same length as by).

priority

integerish(1)
how likely the commodity is produced by the land-use type(s) in response to other land-use types.

exclude

character(.)
land-use types from which the commodity is excluded to be produced. Overrides the other arguments if set.

unit

character(1)
the unit of amount.

Details

A commodity can be any good or service and it can be produced by several land-use types. However, often distinct land-use types produce different amounts of the commodities and they produce them with a different likeliness. This likeliness is specified as land-use type specific priority value.