Modify parts of a model

Modify parts of a model

Usage

useModifier(module = "crop_rotation", ...)

Arguments

module

character(1)
modifications of CLUMondo typically come with a set of files that have to be adapted and these are summarised into modules. By selecting a module you thus make sure that all adaptions are done consistently.

...

module specific objects for that modification. See documentation of the respective module.

Details

Availble modules

mod_crop_rotation: When using this module, CLUMondo will be run separately for each year by default and re-calculate 'locspec' after each step.